Refactor footer to use flexbox
Summary:
Note that this diff has a hardcoded path in .arcconfig because of some
weird arc issue. Don't merge until that's resolved.

Before, the footer used the foundation grid system. However, this was
problematic because of the specific behavior we wanted on smaller screen
sizes. For example, we needed to hardcode two columns breaking to the
next row in HTML and CSS.

I gave this refactor a shot by using flexbox and following the
conventions from http://cssguidelin.es/. Consider this an experiment in
seeing of those worked out.

Next steps are to make the footer stick to the bottom of the page even
when the page doesn't have a full page height's of content. This is also
easily done with flexbox.

Depends on D109 since scss_lint doesn't like this commit very much.

Desktop
-----
{F7017 size=full}

Tablet
-----
{F7016 size=full}

Mobile
-----
{F7015 size=full}

Test Plan:
Go to the home page and ensure that the footer looks good on screen
sizes from 320px to 1024px.

Reviewers: nnarayen, aleks

Subscribers: bigblue

Differential Revision: http://phab.calblueprint.org/D106