[Projects 2] Add admin-specific layout, CSS, and JS files
Summary:
This is because a good amount of CSS and JS on the admin side won't be
used on the public-facing side, so we can cut some page load time in
that regard.

The real reason I'm doing this is so that we can use the admin side as a
sandbox of sorts where we can experiment with libraries and
modifications without affecting the public side.

In other words, I'd like to play around with React on the admin side.

The CSS I added fixes the footer not sticking to the bottom when the
page content is not long enough for a whole page based on
https://philipwalton.github.io/solved-by-flexbox/demos/sticky-footer/

Depends on D110

Test Plan:
- Visit http://localhost:3000/admins/settings and verify that the footer
  is bottomed out correctly even though the content is not an entire
  page height.
- Visit http://localhost:3000/contact_forms/new , delete the content
  with Chrome Devtools, and verify that the footer still does not bottom
  out.
- Visit http://localhost:3000/admins/projects and verify that "Hello
  world" shows up.

Reviewers: nnarayen

Reviewed By: nnarayen

Subscribers: bigblue, aleks

Projects: #calblueprint.org

Differential Revision: http://phab.calblueprint.org/D111