[Projects 6] Implement project cards
Summary:
Before, the projects on `/projects` were hard-coded into the HTML. In
this diff, they are loaded dynamically from the database.

I also restyled the cards to just show basic info with a link to the
case study for the project.

Card animation
-----
Ignore the weirdness in the beginning.
{F8381 size=full}

Small screen
-----
{F8375 width=320}

Medium screen
-----
{F8376 width=500}

Large screen
-----
{F8377 width=500}

Next steps are to implement the following:
- Uploading of banner image on the admin side.
- The project show page
- Ordering of projects (using `acts_as_list`)
- Editing of project case study text on admin side.

Depends on D115

Test Plan:
- Make sure projects are seeded with `rake update:projects`.
- Visit http://localhost:3000/projects and verify that each project in
  the db shows up.

Reviewers: nnarayen

Reviewed By: nnarayen

Subscribers: bigblue, aleks

Projects: #calblueprint.org

Differential Revision: http://phab.calblueprint.org/D116