[Projects 8] Move text-align:center to body
Summary:
Before, we used `text-align: center` on all of our text tags. This
meant that HTML like this:

```
.some-div
  p Some text
```

Couldn't be left-aligned unless you wrote a rather nasty CSS rule that
looked like

```
.some-div {
  // This p selector is needed here because of specificity
  p {
    text-align: left;
  }
}
```

This commit moves the `text-align: center` to the `body` so that we get
center-aligning by default but can style things like:

```
.some-div {
  text-align: left;
}
```

And have everything inside be left-aligned as expected.

Test Plan: - Check all the public-facing pages and make sure that nothing is broken

Reviewers: nnarayen

Reviewed By: nnarayen

Subscribers: bigblue, aleks

Projects: #calblueprint.org

Differential Revision: http://phab.calblueprint.org/D118